requirements.txt
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
vertexai==0.0.1
